Add files via upload
We are going to make a "Hangman Letter" game. 
The computer will pick a word.
The player can guess it letter by letter or run out of chances.
But if they make too many wrong guesses, they lose.
If the player makes the right guesses he wins.
Cleaner interface and option to quit the game.